+++ lib/s_packages.dart
@@ -41,4 +41,6 @@ export 'states_rebuilder_extended/states_rebuilder_extended.dart' hide Animate;
 export 'ticker_free_circular_progress_indicator/ticker_free_circular_progress_indicator.dart';
 export 'week_calendar/week_calendar.dart';
 
+export 'package:nb_utils/nb_utils.dart';
+
 export 's_packages_extra1.dart';